Comparison - A* vs Dijkstra's v1
Question 2 - The creator of the puzzle has been told that the A* algorithm is more efficient at finding the shortest path because it uses heuristics.

Compare the performance of Dijkstra’s algorithm and the A* search algorithm, referring to heuristics, to find the shortest path to the problem by implementing both algorithms programmatically and comparing the solutions generated in Mark-down.

Refer to the complexity of the algorithms and compare the actual time it takes for the solutions to be processed.